refactor(dtt): v0.0.10 polish — modularise datetime, gate serde, tighten lints
Five-part architectural polish landing within the v0.0.10 release branch.
Public API unchanged; the `serde` feature defaults on, so existing
downstream behaviour is preserved.

1. Modularise `src/datetime.rs` (2492 lines) →
   `src/datetime/{mod,builder,validate,tests}.rs`
   - mod.rs:      core type, TZ_OFFSETS, primary impls, trait impls,
                  helpers (~1645 lines)
   - builder.rs:  DateTimeBuilder (~159 lines)
   - validate.rs: `impl DateTime { is_valid_* }` validators (~95 lines)
   - tests.rs:    the inline unit-test module (~623 lines)
   Re-exports keep `dtt::datetime::{DateTime,DateTimeBuilder}` working.

2. Migrate clippy enforcement to [lints.clippy] in Cargo.toml for
   universal allowances; library-only strict lints (pedantic, nursery,
   cargo, unwrap_used, expect_used, panic, result_unit_err,
   clone_on_ref_ptr) remain as inner #![deny(...)] in src/lib.rs so
   they don't fire on tests/benches/examples that legitimately use
   unwrap. Dead [package.metadata.clippy] block removed.

3. Make `serde` optional via `[features] default = ["serde"]`.
   - serde + serde_json now `optional = true`; pulled in by the
     feature
   - time's `serde` feature is gated through dtt's `serde` feature
   - Serialize/Deserialize impls on DateTime, DateTimeBuilder, and
     DateTimeError are #[cfg(feature = "serde")]
   - AppError::SerializationError(serde_json::Error) is cfg-gated
   - The `dtt` example (uses serde_json) declares
     `required-features = ["serde"]`
   - Integration tests using serde_json are individually cfg-gated;
     the all-serde `serialization_tests` mod in test_error.rs is
     cfg-gated as a whole
   - Verified `cargo build --no-default-features --all-targets` clean

4. `clippy::unwrap_used`/`expect_used`/`panic`/`result_unit_err`/
   `clone_on_ref_ptr` denied at the lib crate level (was already so
   for `src/datetime.rs`; now uniform across `src/`).

5. README doctests confirmed passing (56/56). `lib.rs` already had
   `#![doc = include_str!("../README.md")]` so the README is the
   crate-level doc.

Drive-by CI fix:
- build.rs: `eprintln!("…{}", v)` → `eprintln!("…{v}")` to satisfy
  clippy::uninlined_format_args under the now-active lint scope.

Verification: 209 unit/integration tests pass, 56 doctests pass,
cargo clippy --all-targets -- -D warnings clean,
cargo build --no-default-features --all-targets clean,
cargo build --release clean, make sbom produces dtt-sbom.cdx.json.

---
THE ARCHITECT ᛫ Sebastien Rousseau ᛫ https://sebastienrousseau.com
THE ENGINE ᛞ EUXIS ᛫ Enterprise Unified Execution Intelligence System ᛫ https://euxis.co
