+++ Cargo.toml
@@ -57,6 +57,16 @@ name = "benchmark"
 path = "benches/criterion.rs"
 harness = false
 
+# -----------------------------------------------------------------------------
+# Examples
+# -----------------------------------------------------------------------------
+# The `dtt` example exercises serde round-tripping; it can only build when
+# the `serde` feature is enabled. The other examples are feature-free.
+[[example]]
+name = "dtt"
+path = "examples/dtt.rs"
+required-features = ["serde"]
+
 [profile.bench]
 debug = true
 
@@ -67,10 +77,13 @@ debug = true
 # `pastey` is the maintained fork of the now-unmaintained `paste` crate
 # (RUSTSEC-2024-0436). Drop-in compatible.
 pastey = "0.2.1"
-serde = { version = "1.0.228", features = ["derive"] }
-serde_json = "1.0.140"
+# serde is opt-in via the `serde` feature (enabled by default). Disable
+# default-features to drop ~6 transitive dependencies.
+serde = { version = "1.0.228", features = ["derive"], optional = true }
+serde_json = { version = "1.0.140", optional = true }
 # 0.3.47 fixes RUSTSEC time stack-exhaustion DoS (medium severity).
-time = { version = "0.3.47", features = ["serde", "formatting", "parsing"] }
+# `time`'s `serde` feature is only needed when our `serde` feature is on.
+time = { version = "0.3.47", default-features = false, features = ["formatting", "parsing"] }
 thiserror = "2.0.18"
 
 # -----------------------------------------------------------------------------
@@ -94,7 +107,10 @@ version_check = "0.9.5"
 # Features
 # -----------------------------------------------------------------------------
 [features]
-default = []
+# Default keeps the v0.0.9 behaviour (serde on, RFC 3339 round-trip).
+default = ["serde"]
+# Optional `serde`/`serde_json` support. Drops ~6 transitive crates when off.
+serde = ["dep:serde", "dep:serde_json", "time/serde"]
 bench = []
 
 # -----------------------------------------------------------------------------
@@ -151,13 +167,16 @@ unused_macro_rules = "deny"
 unused_qualifications = "deny"
 variant_size_differences = "deny"
 
-[package.metadata.clippy]
-warn-lints = [
-    "clippy::all",
-    "clippy::pedantic",
-    "clippy::cargo",
-    "clippy::nursery",
-]
+# Cargo-managed clippy enforcement applies to ALL targets in the package
+# (lib, bin, tests, benches, examples). Keep only universal allowances
+# here; library-only strict lints (`pedantic`, `nursery`, `cargo`,
+# `unwrap_used`, etc.) live as inner `#![deny(...)]` attributes in
+# `src/lib.rs` so they don't fire on tests that legitimately use unwrap.
+[lints.clippy]
+module_name_repetitions = "allow"
+# `multiple_crate_versions` fires on dev-deps the user can't control
+# (criterion brings in older itertools etc.).
+multiple_crate_versions = "allow"
 
 [profile.dev]
 codegen-units = 256
