+++ src/datetime/mod.rs
@@ -36,17 +36,10 @@
 //! }
 //! ```
 
-#![deny(
-    missing_docs,
-    clippy::unwrap_used,
-    clippy::expect_used,
-    clippy::panic,
-    clippy::result_unit_err,
-    clippy::clone_on_ref_ptr
-)]
-#![warn(clippy::pedantic, clippy::nursery, clippy::cargo)]
+// Lints are enforced via [lints.clippy] and [lints.rust] in Cargo.toml.
 
 use crate::error::DateTimeError;
+#[cfg(feature = "serde")]
 use serde::{Deserialize, Deserializer, Serialize, Serializer};
 use std::{
     cmp::Ordering,
@@ -62,26 +55,36 @@ use time::{
     PrimitiveDateTime, Time, UtcOffset, Weekday,
 };
 
+// Submodules. `DateTimeBuilder` is re-exported below so callers can keep
+// using `dtt::datetime::DateTimeBuilder`. The `validate` module hosts an
+// additional `impl DateTime { ... }` block and needs no re-export.
+mod builder;
+mod validate;
+#[cfg(test)]
+mod tests;
+
+pub use builder::DateTimeBuilder;
+
 /// Maximum valid hour value (0-23)
-const MAX_HOUR: u8 = 23;
+pub(super) const MAX_HOUR: u8 = 23;
 
 /// Maximum valid minute/second value (0-59)
-const MAX_MIN_SEC: u8 = 59;
+pub(super) const MAX_MIN_SEC: u8 = 59;
 
 /// Maximum valid day value (1-31)
-const MAX_DAY: u8 = 31;
+pub(super) const MAX_DAY: u8 = 31;
 
 /// Maximum valid month value (1-12)
-const MAX_MONTH: u8 = 12;
+pub(super) const MAX_MONTH: u8 = 12;
 
 /// Maximum valid microsecond value (0-999_999)
-const MAX_MICROSECOND: u32 = 999_999;
+pub(super) const MAX_MICROSECOND: u32 = 999_999;
 
 /// Maximum valid ISO week number (1-53)
-const MAX_ISO_WEEK: u8 = 53;
+pub(super) const MAX_ISO_WEEK: u8 = 53;
 
 /// Maximum valid ordinal day (1-366)
-const MAX_ORDINAL_DAY: u16 = 366;
+pub(super) const MAX_ORDINAL_DAY: u16 = 366;
 
 /// Represents a date and time with timezone offset support.
 ///
@@ -109,6 +112,7 @@ pub struct DateTime {
     pub(crate) offset: UtcOffset,
 }
 
+#[cfg(feature = "serde")]
 impl Serialize for DateTime {
     /// Serializes a `DateTime` as a canonical RFC 3339 string. Two
     /// `DateTime` values that compare equal under `PartialEq` always
@@ -123,6 +127,7 @@ impl Serialize for DateTime {
     }
 }
 
+#[cfg(feature = "serde")]
 impl<'de> Deserialize<'de> for DateTime {
     /// Deserializes a `DateTime` from an RFC 3339 string.
     fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
@@ -268,150 +273,6 @@ static TIMEZONE_OFFSETS: LazyLock<
 
     m
 });
-
-// -----------------------------------------------------------------------------
-// Builder Pattern
-// -----------------------------------------------------------------------------
-
-/// A builder for [`DateTime`] objects, allowing more ergonomic creation of
-/// datetimes with customized year, month, day, hour, minute, second, and offset.
-///
-/// # Examples
-///
-/// ```
-/// use dtt::datetime::{DateTime, DateTimeBuilder};
-/// use time::UtcOffset;
-///
-/// let builder = DateTimeBuilder::new()
-///     .year(2024)
-///     .month(1)
-///     .day(1)
-///     .hour(12)
-///     .minute(30)
-///     .second(45)
-///     .offset(UtcOffset::UTC);
-///
-/// let dt = builder.build();
-/// assert!(dt.is_ok());
-///
-/// let dt_unwrapped = dt.unwrap();
-/// assert_eq!(dt_unwrapped.year(), 2024);
-/// assert_eq!(dt_unwrapped.month().to_string(), "January");
-/// assert_eq!(dt_unwrapped.day(), 1);
-/// assert_eq!(dt_unwrapped.hour(), 12);
-/// assert_eq!(dt_unwrapped.minute(), 30);
-/// assert_eq!(dt_unwrapped.second(), 45);
-/// assert_eq!(dt_unwrapped.offset(), UtcOffset::UTC);
-/// ```
-#[derive(
-    Clone, Copy, Debug, PartialEq, Eq, Hash, Serialize, Deserialize,
-)]
-pub struct DateTimeBuilder {
-    /// Calendar year, e.g. 2024.
-    year: i32,
-    /// Month (1-12).
-    month: u8,
-    /// Day of the month (1-31, depends on month).
-    day: u8,
-    /// Hour of the day (0-23).
-    hour: u8,
-    /// Minute of the hour (0-59).
-    minute: u8,
-    /// Second of the minute (0-59).
-    second: u8,
-    /// The time zone offset from UTC.
-    offset: UtcOffset,
-}
-
-impl Default for DateTimeBuilder {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-impl DateTimeBuilder {
-    /// Creates a new `DateTimeBuilder` with default values set to
-    /// midnight, January 1, 1970 (UTC).
-    #[must_use]
-    pub const fn new() -> Self {
-        Self {
-            year: 1970,
-            month: 1,
-            day: 1,
-            hour: 0,
-            minute: 0,
-            second: 0,
-            offset: UtcOffset::UTC,
-        }
-    }
-
-    /// Sets the year component.
-    #[must_use]
-    pub const fn year(mut self, year: i32) -> Self {
-        self.year = year;
-        self
-    }
-
-    /// Sets the month component.
-    #[must_use]
-    pub const fn month(mut self, month: u8) -> Self {
-        self.month = month;
-        self
-    }
-
-    /// Sets the day component.
-    #[must_use]
-    pub const fn day(mut self, day: u8) -> Self {
-        self.day = day;
-        self
-    }
-
-    /// Sets the hour component.
-    #[must_use]
-    pub const fn hour(mut self, hour: u8) -> Self {
-        self.hour = hour;
-        self
-    }
-
-    /// Sets the minute component.
-    #[must_use]
-    pub const fn minute(mut self, minute: u8) -> Self {
-        self.minute = minute;
-        self
-    }
-
-    /// Sets the second component.
-    #[must_use]
-    pub const fn second(mut self, second: u8) -> Self {
-        self.second = second;
-        self
-    }
-
-    /// Sets the time zone offset component.
-    #[must_use]
-    pub const fn offset(mut self, offset: UtcOffset) -> Self {
-        self.offset = offset;
-        self
-    }
-
-    /// Builds the final [`DateTime`] from the builder state.
-    ///
-    /// # Errors
-    ///
-    /// Returns a `DateTimeError` if any of the date components are invalid
-    /// (e.g., `month = 13` or `day = 32`).
-    pub fn build(&self) -> Result<DateTime, DateTimeError> {
-        DateTime::from_components(
-            self.year,
-            self.month,
-            self.day,
-            self.hour,
-            self.minute,
-            self.second,
-            self.offset,
-        )
-    }
-}
 
 // -----------------------------------------------------------------------------
 // Core Implementations
@@ -1608,103 +1469,6 @@ impl DateTime {
     }
 }
 
-// -----------------------------------------------------------------------------
-// Validation Methods
-// -----------------------------------------------------------------------------
-
-impl DateTime {
-    /// Validates whether a string represents a valid day of the month.
-    #[must_use]
-    pub fn is_valid_day(day: &str) -> bool {
-        day.parse::<u8>()
-            .map(|d| (1..=MAX_DAY).contains(&d))
-            .unwrap_or(false)
-    }
-
-    /// Validates whether a string represents a valid hour.
-    #[must_use]
-    pub fn is_valid_hour(hour: &str) -> bool {
-        hour.parse::<u8>().map(|h| h <= MAX_HOUR).unwrap_or(false)
-    }
-
-    /// Validates whether a string represents a valid minute.
-    #[must_use]
-    pub fn is_valid_minute(minute: &str) -> bool {
-        minute
-            .parse::<u8>()
-            .map(|m| m <= MAX_MIN_SEC)
-            .unwrap_or(false)
-    }
-
-    /// Validates whether a string represents a valid second.
-    #[must_use]
-    pub fn is_valid_second(second: &str) -> bool {
-        second
-            .parse::<u8>()
-            .map(|s| s <= MAX_MIN_SEC)
-            .unwrap_or(false)
-    }
-
-    /// Validates whether a string represents a valid month.
-    #[must_use]
-    pub fn is_valid_month(month: &str) -> bool {
-        month
-            .parse::<u8>()
-            .map(|m| (1..=MAX_MONTH).contains(&m))
-            .unwrap_or(false)
-    }
-
-    /// Validates whether a string represents a valid year within the
-    /// supported `time::Date` range of `-9999..=9999`.
-    ///
-    /// This bound matches what `DateTime::from_components` can actually
-    /// construct, so the validator and the builder agree.
-    #[must_use]
-    pub fn is_valid_year(year: &str) -> bool {
-        year.parse::<i32>()
-            .map(|y| (-9999..=9999).contains(&y))
-            .unwrap_or(false)
-    }
-
-    /// Validates whether a string represents a valid microsecond.
-    #[must_use]
-    pub fn is_valid_microsecond(microsecond: &str) -> bool {
-        microsecond
-            .parse::<u32>()
-            .map(|us| us <= MAX_MICROSECOND)
-            .unwrap_or(false)
-    }
-
-    /// Validates whether a string represents a valid ordinal day of the year.
-    #[must_use]
-    pub fn is_valid_ordinal(ordinal: &str) -> bool {
-        ordinal
-            .parse::<u16>()
-            .map(|o| (1..=MAX_ORDINAL_DAY).contains(&o))
-            .unwrap_or(false)
-    }
-
-    /// Validates whether a string represents a valid ISO week number.
-    #[must_use]
-    pub fn is_valid_iso_week(week: &str) -> bool {
-        week.parse::<u8>()
-            .map(|w| (1..=MAX_ISO_WEEK).contains(&w))
-            .unwrap_or(false)
-    }
-
-    /// Validates whether a string represents a valid time in `HH:MM:SS` format.
-    #[must_use]
-    pub fn is_valid_time(time: &str) -> bool {
-        let parts: Vec<&str> = time.split(':').collect();
-        if parts.len() != 3 {
-            return false;
-        }
-
-        Self::is_valid_hour(parts[0])
-            && Self::is_valid_minute(parts[1])
-            && Self::is_valid_second(parts[2])
-    }
-}
 
 // -----------------------------------------------------------------------------
 // Standard Trait Implementations
@@ -1891,622 +1655,3 @@ pub const fn is_leap_year(year: i32) -> bool {
     (year % 4 == 0 && year % 100 != 0) || (year % 400 == 0)
 }
 
-// -----------------------------------------------------------------------------
-// Tests
-// -----------------------------------------------------------------------------
-
-#[cfg(test)]
-#[allow(clippy::unwrap_used, clippy::expect_used)]
-mod tests {
-    use super::*;
-    use std::str::FromStr;
-
-    #[test]
-    fn test_new() {
-        let dt = DateTime::new();
-        assert_eq!(dt.offset(), UtcOffset::UTC);
-    }
-
-    #[test]
-    fn test_new_with_tz() {
-        let est_dt = DateTime::new_with_tz("EST_USA").unwrap();
-        assert_eq!(est_dt.offset().whole_hours(), -5);
-
-        let invalid = DateTime::new_with_tz("INVALID");
-        assert!(matches!(invalid, Err(DateTimeError::InvalidTimezone)));
-    }
-
-    #[test]
-    fn test_new_with_custom_offset() {
-        let dt = DateTime::new_with_custom_offset(5, 30).unwrap();
-        assert_eq!(dt.offset().whole_hours(), 5);
-        assert_eq!(dt.offset().minutes_past_hour(), 30);
-
-        // Test invalid offsets
-        let too_large_hours = DateTime::new_with_custom_offset(24, 0);
-        assert!(too_large_hours.is_err());
-        let too_large_minutes = DateTime::new_with_custom_offset(0, 60);
-        assert!(too_large_minutes.is_err());
-    }
-
-    #[test]
-    fn test_from_components() {
-        let dt_val = DateTime::from_components(
-            2024,
-            1,
-            1,
-            12,
-            0,
-            0,
-            UtcOffset::UTC,
-        )
-        .unwrap();
-        assert_eq!(dt_val.year(), 2024);
-        assert_eq!(dt_val.month(), Month::January);
-        assert_eq!(dt_val.day(), 1);
-        assert_eq!(dt_val.hour(), 12);
-        assert_eq!(dt_val.minute(), 0);
-        assert_eq!(dt_val.second(), 0);
-
-        // Test invalid dates
-        let invalid_month = DateTime::from_components(
-            2024,
-            13,
-            1,
-            0,
-            0,
-            0,
-            UtcOffset::UTC,
-        );
-        assert!(invalid_month.is_err());
-
-        let invalid_day = DateTime::from_components(
-            2024,
-            2,
-            30,
-            0,
-            0,
-            0,
-            UtcOffset::UTC,
-        );
-        assert!(invalid_day.is_err());
-    }
-
-    #[test]
-    fn test_parse() {
-        // Test RFC 3339 format
-        let dt = DateTime::parse("2024-01-01T12:00:00Z");
-        assert!(dt.is_ok());
-
-        // Test ISO 8601 date
-        let dt_val = DateTime::parse("2024-01-01").unwrap();
-        assert_eq!(dt_val.hour(), 0);
-        assert_eq!(dt_val.minute(), 0);
-
-        // Test invalid formats
-        let invalid1 = DateTime::parse("invalid");
-        assert!(invalid1.is_err());
-        let invalid2 = DateTime::parse("2024-13-01");
-        assert!(invalid2.is_err());
-    }
-
-    #[test]
-    fn test_format() {
-        let dt = DateTime::new();
-        let maybe_formatted = dt.format("[year]-[month]-[day]");
-        assert!(maybe_formatted.is_ok());
-
-        let invalid_format = dt.format("[invalid]");
-        assert!(invalid_format.is_err());
-    }
-
-    #[test]
-    fn test_timezone_conversion() {
-        let utc = DateTime::new();
-        let est_val = utc.convert_to_tz("EST_USA").unwrap();
-        assert_eq!(est_val.offset().whole_hours(), -5);
-
-        let invalid = utc.convert_to_tz("INVALID");
-        assert!(invalid.is_err());
-    }
-
-    #[test]
-    fn test_arithmetic() {
-        let dt = DateTime::new();
-
-        // Test adding days
-        let future = dt.add_days(7);
-        assert!(future.is_ok());
-
-        // Test subtracting days (negative)
-        let past = dt.add_days(-7);
-        assert!(past.is_ok());
-
-        // Test adding months
-        let next_month = dt.add_months(1);
-        assert!(next_month.is_ok());
-
-        // Test month edge cases
-        let jan31_dt = DateTime::from_components(
-            2024,
-            1,
-            31,
-            0,
-            0,
-            0,
-            UtcOffset::UTC,
-        )
-        .unwrap();
-        let feb_dt = jan31_dt.add_months(1).unwrap();
-        // 2024 is a leap year => Feb has 29 days
-        assert_eq!(feb_dt.day(), 29);
-    }
-
-    #[test]
-    fn test_leap_year() {
-        assert!(is_leap_year(2024));
-        assert!(!is_leap_year(2023));
-        assert!(is_leap_year(2000));
-        assert!(!is_leap_year(1900));
-    }
-
-    #[test]
-    fn test_validation() {
-        // Test day validation
-        assert!(DateTime::is_valid_day("1"));
-        assert!(DateTime::is_valid_day("31"));
-        assert!(!DateTime::is_valid_day("0"));
-        assert!(!DateTime::is_valid_day("32"));
-        assert!(!DateTime::is_valid_day("abc"));
-
-        // Test hour validation
-        assert!(DateTime::is_valid_hour("0"));
-        assert!(DateTime::is_valid_hour("23"));
-        assert!(!DateTime::is_valid_hour("24"));
-
-        // Test minute validation
-        assert!(DateTime::is_valid_minute("0"));
-        assert!(DateTime::is_valid_minute("59"));
-        assert!(!DateTime::is_valid_minute("60"));
-
-        // Test time string validation
-        assert!(DateTime::is_valid_time("00:00:00"));
-        assert!(DateTime::is_valid_time("23:59:59"));
-        assert!(!DateTime::is_valid_time("24:00:00"));
-        assert!(!DateTime::is_valid_time("23:60:00"));
-        assert!(!DateTime::is_valid_time("23:59:60"));
-
-        // Year validation must agree with what `from_components` accepts:
-        // the `time::Date` range is -9999..=9999.
-        assert!(DateTime::is_valid_year("0"));
-        assert!(DateTime::is_valid_year("2024"));
-        assert!(DateTime::is_valid_year("9999"));
-        assert!(DateTime::is_valid_year("-9999"));
-        assert!(!DateTime::is_valid_year("10000"));
-        assert!(!DateTime::is_valid_year("-10000"));
-        assert!(!DateTime::is_valid_year("99999999"));
-        assert!(!DateTime::is_valid_year("abc"));
-    }
-
-    #[test]
-    fn test_iso_year_at_boundary() {
-        // 2022-01-01 is a Saturday and falls in ISO week 52 of 2021,
-        // so its ISO year (2021) differs from its calendar year (2022).
-        let dt = DateTime::from_components(
-            2022,
-            1,
-            1,
-            0,
-            0,
-            0,
-            UtcOffset::UTC,
-        )
-        .unwrap();
-        assert_eq!(dt.year(), 2022);
-        assert_eq!(dt.iso_week(), 52);
-        assert_eq!(dt.iso_year(), 2021);
-
-        // Mid-year: ISO year matches calendar year.
-        let mid = DateTime::from_components(
-            2024,
-            6,
-            15,
-            0,
-            0,
-            0,
-            UtcOffset::UTC,
-        )
-        .unwrap();
-        assert_eq!(mid.year(), mid.iso_year());
-    }
-
-    #[test]
-    fn test_range_operations() {
-        let dt_val = DateTime::from_components(
-            2024,
-            1,
-            15,
-            12,
-            0,
-            0,
-            UtcOffset::UTC,
-        )
-        .unwrap();
-        // Test week ranges
-        let ws = dt_val.start_of_week().unwrap();
-        assert_eq!(ws.weekday(), Weekday::Monday);
-
-        let we = dt_val.end_of_week().unwrap();
-        assert_eq!(we.weekday(), Weekday::Sunday);
-
-        // Test month ranges
-        let ms = dt_val.start_of_month().unwrap();
-        assert_eq!(ms.day(), 1);
-
-        let me = dt_val.end_of_month().unwrap();
-        assert_eq!(me.day(), 31);
-
-        // Test year ranges
-        let ys = dt_val.start_of_year().unwrap();
-        assert_eq!(ys.month(), Month::January);
-        assert_eq!(ys.day(), 1);
-
-        let ye = dt_val.end_of_year().unwrap();
-        assert_eq!(ye.month(), Month::December);
-        assert_eq!(ye.day(), 31);
-    }
-
-    #[test]
-    fn test_ordering() {
-        let dt1 = DateTime::from_components(
-            2024,
-            1,
-            1,
-            0,
-            0,
-            0,
-            UtcOffset::UTC,
-        );
-        let dt2 = DateTime::from_components(
-            2024,
-            1,
-            2,
-            0,
-            0,
-            0,
-            UtcOffset::UTC,
-        );
-
-        let a = dt1.unwrap();
-        let b = dt2.unwrap();
-        assert!(a < b);
-        assert!(b > a);
-        assert_ne!(a, b);
-    }
-
-    #[test]
-    fn test_duration() {
-        let dt1 = DateTime::from_components(
-            2024,
-            1,
-            1,
-            0,
-            0,
-            0,
-            UtcOffset::UTC,
-        );
-        let dt2 = DateTime::from_components(
-            2024,
-            1,
-            2,
-            0,
-            0,
-            0,
-            UtcOffset::UTC,
-        );
-
-        let a = dt1.unwrap();
-        let b = dt2.unwrap();
-        let duration = b.duration_since(&a);
-        assert_eq!(duration.whole_days(), 1);
-    }
-
-    #[test]
-    fn test_from_str() {
-        let dt = DateTime::from_str("2024-01-01T00:00:00Z");
-        assert!(dt.is_ok());
-        let invalid = DateTime::from_str("invalid");
-        assert!(invalid.is_err());
-    }
-
-    #[test]
-    fn test_display() {
-        let dt_val = DateTime::from_components(
-            2024,
-            1,
-            1,
-            0,
-            0,
-            0,
-            UtcOffset::UTC,
-        )
-        .unwrap();
-        assert_eq!(dt_val.to_string(), "2024-01-01T00:00:00Z");
-    }
-
-    #[test]
-    fn test_hash() {
-        use std::collections::HashSet;
-        let dt1 = DateTime::from_components(
-            2024,
-            1,
-            1,
-            0,
-            0,
-            0,
-            UtcOffset::UTC,
-        );
-        let dt2 = DateTime::from_components(
-            2024,
-            1,
-            1,
-            0,
-            0,
-            0,
-            UtcOffset::UTC,
-        );
-        let a = dt1.unwrap();
-        let b = dt2.unwrap();
-        let mut set = HashSet::new();
-        assert!(
-            set.insert(a),
-            "The set should not have contained `a` before"
-        );
-        assert!(set.contains(&b));
-    }
-
-    #[test]
-    fn test_builder_pattern() {
-        let builder = DateTimeBuilder::new()
-            .year(2024)
-            .month(1)
-            .day(1)
-            .hour(12)
-            .minute(30)
-            .second(45)
-            .offset(UtcOffset::UTC);
-
-        let value = builder.build().unwrap();
-        assert_eq!(value.year(), 2024);
-        assert_eq!(value.month(), Month::January);
-        assert_eq!(value.day(), 1);
-        assert_eq!(value.hour(), 12);
-        assert_eq!(value.minute(), 30);
-        assert_eq!(value.second(), 45);
-    }
-
-    // -------------------------------------------------------------------------
-    // Error-path coverage. These tests pin every fallible production branch
-    // so that the coverage report stays at (or near) 100% and a future
-    // refactor cannot silently turn an error into a panic.
-    // -------------------------------------------------------------------------
-
-    #[test]
-    fn test_set_time_invalid_returns_err() {
-        let dt = DateTime::new();
-        // Hour out of range — exercises the `Time::from_hms` Err arm.
-        assert!(matches!(
-            dt.set_time(24, 0, 0),
-            Err(DateTimeError::InvalidTime)
-        ));
-        assert!(matches!(
-            dt.set_time(0, 60, 0),
-            Err(DateTimeError::InvalidTime)
-        ));
-        assert!(matches!(
-            dt.set_time(0, 0, 60),
-            Err(DateTimeError::InvalidTime)
-        ));
-    }
-
-    #[test]
-    fn test_set_date_invalid_month_returns_err() {
-        let dt = DateTime::new();
-        // Month 13 — exercises the `Month::try_from` Err arm in `set_date`.
-        assert!(matches!(
-            dt.set_date(2024, 13, 1),
-            Err(DateTimeError::InvalidDate)
-        ));
-        // Day 30 in February — exercises the `Date::from_calendar_date` Err arm.
-        assert!(matches!(
-            dt.set_date(2024, 2, 30),
-            Err(DateTimeError::InvalidDate)
-        ));
-    }
-
-    #[test]
-    fn test_parse_custom_format_invalid_input_returns_err() {
-        // Format is valid, input does not match it.
-        let result = DateTime::parse_custom_format(
-            "not-a-date",
-            "[year]-[month]-[day]",
-        );
-        assert!(matches!(result, Err(DateTimeError::InvalidFormat)));
-    }
-
-    #[test]
-    fn test_add_months_overflow_returns_err() {
-        let dt = DateTime::from_components(
-            9999,
-            12,
-            1,
-            0,
-            0,
-            0,
-            UtcOffset::UTC,
-        )
-        .unwrap();
-        // i32::MAX months will overflow the year*12 + months calculation,
-        // exercising the `checked_*` Err arm.
-        assert!(matches!(
-            dt.add_months(i32::MAX),
-            Err(DateTimeError::InvalidDate)
-        ));
-        assert!(matches!(
-            dt.add_months(i32::MIN),
-            Err(DateTimeError::InvalidDate)
-        ));
-    }
-
-    #[test]
-    fn test_add_months_year_out_of_range_returns_err() {
-        let dt = DateTime::from_components(
-            9999,
-            12,
-            1,
-            0,
-            0,
-            0,
-            UtcOffset::UTC,
-        )
-        .unwrap();
-        // 12 months past Dec 9999 lands in year 10000, which `time::Date`
-        // refuses by default. Exercises the final `Date::from_calendar_date`
-        // Err arm in `add_months`.
-        assert!(matches!(
-            dt.add_months(12),
-            Err(DateTimeError::InvalidDate)
-        ));
-    }
-
-    #[test]
-    fn test_add_years_overflow_returns_err() {
-        let dt = DateTime::new();
-        // `i32::MAX` years overflows the `checked_add` in `add_years`.
-        assert!(matches!(
-            dt.add_years(i32::MAX),
-            Err(DateTimeError::InvalidDate)
-        ));
-    }
-
-    #[test]
-    fn test_add_years_out_of_range_returns_err() {
-        let dt = DateTime::from_components(
-            9999,
-            6,
-            15,
-            0,
-            0,
-            0,
-            UtcOffset::UTC,
-        )
-        .unwrap();
-        // year 10000 is outside `time::Date::MAX`.
-        assert!(matches!(
-            dt.add_years(1),
-            Err(DateTimeError::InvalidDate)
-        ));
-    }
-
-    #[test]
-    fn test_is_valid_time_wrong_part_count() {
-        // Two parts — exercises the `parts.len() != 3` early return.
-        assert!(!DateTime::is_valid_time("12:00"));
-        // Four parts — same path.
-        assert!(!DateTime::is_valid_time("12:00:00:00"));
-        // Empty.
-        assert!(!DateTime::is_valid_time(""));
-        // Just colons.
-        assert!(!DateTime::is_valid_time("::"));
-    }
-
-    #[test]
-    fn test_serde_round_trip() {
-        let dt = DateTime::from_components(
-            2024,
-            6,
-            15,
-            12,
-            30,
-            45,
-            UtcOffset::UTC,
-        )
-        .unwrap();
-        let json = serde_json::to_string(&dt).unwrap();
-        let parsed: DateTime = serde_json::from_str(&json).unwrap();
-        assert_eq!(dt, parsed);
-    }
-
-    #[test]
-    fn test_serde_deserialize_rejects_invalid_string() {
-        // Exercises the deserialize Err arm by feeding garbage.
-        let result: Result<DateTime, _> =
-            serde_json::from_str("\"not-a-datetime\"");
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn test_format_time_in_timezone_invalid_tz_returns_err() {
-        let dt = DateTime::new();
-        let result =
-            dt.format_time_in_timezone("NOPE_ZONE", "[hour]:[minute]");
-        assert!(matches!(result, Err(DateTimeError::InvalidTimezone)));
-    }
-
-    #[test]
-    fn test_set_time_ok_path() {
-        let dt = DateTime::from_components(
-            2024,
-            6,
-            15,
-            0,
-            0,
-            0,
-            UtcOffset::UTC,
-        )
-        .unwrap();
-        let updated = dt.set_time(10, 30, 45).unwrap();
-        assert_eq!(updated.hour(), 10);
-        assert_eq!(updated.minute(), 30);
-        assert_eq!(updated.second(), 45);
-        // Date and offset preserved.
-        assert_eq!(updated.year(), 2024);
-        assert_eq!(updated.month(), Month::June);
-        assert_eq!(updated.day(), 15);
-        assert_eq!(updated.offset(), UtcOffset::UTC);
-    }
-
-    #[test]
-    fn test_is_valid_iso_week() {
-        assert!(DateTime::is_valid_iso_week("1"));
-        assert!(DateTime::is_valid_iso_week("52"));
-        assert!(DateTime::is_valid_iso_week("53"));
-        assert!(!DateTime::is_valid_iso_week("0"));
-        assert!(!DateTime::is_valid_iso_week("54"));
-        assert!(!DateTime::is_valid_iso_week("not-a-number"));
-    }
-
-    #[test]
-    fn test_datetime_builder_default() {
-        // Exercises `impl Default for DateTimeBuilder { fn default() }`.
-        let builder = DateTimeBuilder::default();
-        let dt = builder.build().unwrap();
-        assert_eq!(dt.year(), 1970);
-        assert_eq!(dt.month(), Month::January);
-        assert_eq!(dt.day(), 1);
-        assert_eq!(dt.offset(), UtcOffset::UTC);
-    }
-
-    #[test]
-    fn test_days_in_month_invalid_month() {
-        // Exercises the catch-all `_ => Err(...)` arm in `days_in_month`.
-        assert!(matches!(
-            days_in_month(2024, 0),
-            Err(DateTimeError::InvalidDate)
-        ));
-        assert!(matches!(
-            days_in_month(2024, 13),
-            Err(DateTimeError::InvalidDate)
-        ));
-    }
-}
